feat: 노선이 삭제되기 이전에 comfirm 창을 띄우기

+++ src/ui/contents-ui/section-manager-ui.js
@@ -95,6 +95,9 @@ export default class SectionManagerUI {
     const deleteButtons = getAllElementsByClass(LINE_DELETE_BUTTON_CLASS);
     Array.prototype.forEach.call(deleteButtons, (deleteButton) => {
       deleteButton.addEventListener("click", (e) => {
+        if (!confirm(DELETE_CONFIRM_MESSAGE)) {
+          return;
+        }
         this.stationINFOManager_.deleteLine(e.target.dataset.name);
         this.updateLinesTable();
       });
